+++ CHANGES.rst
@@ -59,6 +59,7 @@ Development
 - [ui] Fix build by upgrading to underscore-template-loader 1.2.0
 - [mw] Fix access to DEPATISnet due to upstream changes
 - [qa] Speed up testing using ``pytest-xdist``
+- [mw] Improve caching configuration for ``patzilla.access.dpma.dpmaregister``
 
 
 2019-11-01 0.169.3